refactor(nonces): break out into sep module
