+++ docs/api/nonces.rst
@@ -1,0 +1,5 @@
+oneid.nonces
+===========
+
+.. automodule:: oneid.nonces
+   :members: make_nonce, set_nonce_handler, verify_and_burn_nonce
