+++ docs/api/utils.rst
@@ -2,4 +2,4 @@ oneid.utils
 ===========
 
 .. automodule:: oneid.utils
-   :members: make_nonce, verify_and_burn_nonce
+   :members: base64url_encode, base64url_decode
