+++ src/oneid/__init__.py
@@ -2,7 +2,8 @@ from . import keychain
 from . import service
 from . import session
 from . import jwts
+from . import nonces
 from . import utils
 
 
-__all__ = (keychain, service, session, jwts, utils)
+__all__ = (keychain, service, session, jwts, nonces, utils)
